45216: Add a switch to turn on clearing out the /tmp folder after each test is run.

+++ toolset/benchmark/benchmarker.py
@@ -652,15 +652,16 @@ class Benchmarker:
         ##########################################################
         # Remove contents of  /tmp folder
         ##########################################################
-        try:
-          filelist = [ f for f in os.listdir("/tmp") ]
-	  for f in filelist:
-            try:
-	      os.remove("/tmp/" + f)
-            except OSError as err:
-              print "Failed to remove " + str(f) + " from /tmp directory: " + str(err)
-        except OSError:
-          print "Failed to remove contents of /tmp directory."
+        if self.clear_tmp:
+          try:
+            filelist = [ f for f in os.listdir("/tmp") ]
+            for f in filelist:
+              try:
+                os.remove("/tmp/" + f)
+              except OSError as err:
+                print "Failed to remove " + str(f) + " from /tmp directory: " + str(err)
+          except OSError:
+            print "Failed to remove contents of /tmp directory."
 
         ##########################################################
         # Save results thus far into the latest results directory
